Changed the way line-height is treated.
Previously it only controls the extra line height below the text, so if line-height=200%, the line height will be doubled and extra space will be below the text.
Now I made the text vertical position respecting the combination of line-height and vertical-align. So if vertical-align=middle and line-height=200%, the line-height will be doubled and text will be vertically aligned in the middle.